Remove mock subdirectories
