--- internal/app/mapper/mock/mapper.go
@@ -1,11 +1,0 @@
-package mapper
-
-import (
-	"testing"
-
-	"github.com/envoyproxy/xds-relay/internal/app/mapper"
-)
-
-func NewMapper(t *testing.T) mapper.Mapper {
-	return mapper.NewMockMapper(t)
-}
